reorg qc trait and impl
