+++ src/lib.rs
@@ -2,7 +2,7 @@
 // #![warn(missing_docs)] // TODO need rustdoc for stake_table
 
 pub mod circuit;
-pub mod quorum_certificate;
+pub mod qc;
 pub mod stake_table;
 pub mod vdf;
 pub mod vid;
